api/requirements.txt
azure-functions
pydantic
strawberry-graphql
python-fasthtml
azure-cosmos
# uvicorn
# htmx